+++ lib/ToDos.js
@@ -1,61 +1,54 @@
-// const $ = require('jquery');
-// const AllToDos = require('./AllToDos.js');
-//
-// function ToDo(title, body, id, quality) {
-//   this.title = title;
-//   this.body = body;
-//   this.id = id || Date.now();
-//   this.quality = quality || 'swill';
-// }
-//
-//
-// ToDo.prototype.qualityUp = function () {
-//   var quality = this.quality;
-//   var id = this.id;
-//
-//   switch (quality) {
-//     case 'swill':
-//       this.quality = 'plausible';
-//       return AllToDos.store();
-//     case 'plausible':
-//       this.quality = 'genius';
-//       return AllToDos.store();
-//     default:
-//   }
-// };
-//
-// ToDo.prototype.qualityDown = function() {
-//   var quality = this.quality;
-//   switch (quality) {
-//     case 'genius':
-//       this.quality = 'plausible';
-//       return AllToDos.store();
-//     case 'plausible':
-//       this.quality = 'swill';
-//       return AllToDos.store();
-//     default:
-//   }
-//   AllToDos.store();
-// };
-//
-// ToDo.prototype.remove = function(id) {
-//   id = parseInt(id);
-//   debugger;
-//   var array = AllToDos.toDoArray;
-//   AllToDos.toDoArray = array.filter(function (r) {
-//     return r.id !== id;
-//   });
-//   AllToDos.store();
-// };
-//
-// ToDo.prototype.toggleComplete = function(){
-//   // Toggle class of parent container element so that it appears gray / strikethrough text
-//   // Either edit the status of the ToDo item and store in local storage
-//   // Or delete from the ToDo array and store completed ToDo inside of completed ToDo Array
-// };
-//
-// ToDo.prototype.renderCompletedToDos = function(){
-//   //Either grab only ids with status of complete or render todos from todo array
-// };
-//
-// module.exports = ToDo;
+function ToDo(title, body, id, importance, completeness) {
+  this.title = title;
+  this.body = body;
+  this.id = id || Date.now();
+  this.importance = importance || 'Normal';
+  this.completeness = completeness || 'incomplete';
+}
+
+ToDo.prototype.importanceUp = function () {
+  // var imports = importance;
+  var importance = this.importance;
+  var increaseImportance = {
+    "None" : "Low",
+    "Low" : "Normal",
+    "Normal" : "High",
+    "High" : "Critical",
+    "Critical" : "Critical"
+  };
+  this.importance = increaseImportance[importance];
+  return AllToDos.store();
+};
+
+ToDo.prototype.importanceDown = function () {
+  var importance = this.importance;
+  var decreaseImportance = {
+    "Critical":"High",
+    "High" : "Normal",
+    "Normal" : "Low",
+    "Low" : "None",
+    "None" : "None"
+  };
+  this.importance = decreaseImportance[importance];
+  return AllToDos.store();
+};
+
+ToDo.prototype.remove = function(id) {
+  id = parseInt(id);
+  AllToDos.toDoArray = AllToDos.toDoArray.filter(function (r) {
+    return r.id !== id;
+  });
+  AllToDos.store();
+};
+
+ToDo.prototype.toggleComplete = function(){
+  var completeness = this.completeness;
+  var toggleCompleteness = {
+    "complete" : "incomplete",
+    "incomplete" : "complete"
+  };
+  this.completeness = toggleCompleteness[completeness];
+  return AllToDos.store();
+};
+
+module.exports = ToDo;
