+++ test/test.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <link rel="stylesheet" type="text/css" href=".lib/reset.css"/>
+    <link rel="stylesheet" type="text/css" href=".lib/style.css">
+    <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=true;">
+    <link href="https://fonts.googleapis.com/css?family=Open+Sans|Roboto+Slab:700" rel="stylesheet">
+    <title></title>
+  </head>
+  <body>
+    <section class="input">
+        <h1>idea<span id="box">box</span></h1>
+        <input type="text" id="title" placeholder="Title"></input>
+        <textarea id="body" placeholder="Body"></textarea>
+        <button type="submit" id="save">save</button>
+    </section>
+    <section class="ideas">
+        <input type="text" id="search" placeholder="Search"></input>
+        <ul class="card"></ul>
+    </section>
+      <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
+      <script src="test.bundle.js"></script>
+  </body>
+</html>
